Compute ranking of biomarkers by sorting t-test p-values

Compute ranking of biomarkers by sorting t-test p-values

rank_ttest(Y, A, W)

Arguments

Y
(numeric vector) - continuous or binary biomarkers outcome variables

A
(numeric vector) - binary treatment indicator: 1 = treatment, 0 = control

W
(numeric vector, numeric matrix, or numeric data.frame) - matrix of baseline covariates where each column corrspond to one baseline covariate and each row correspond to one observation.

Value

an integer vector containing ranks of biomarkers.

Examples

set.seed(1234)
data(simpleArray)
rank_ttest(Y = simulated_array,
           A = simulated_treatment,
           W = rep(1, length(A)))

#>    [1]    5   26    4    2    3    6   17   60   19    1  855  990  502  508
#>   [15]  244  892  177  325  845  626  279   69  986  163  819  422  164  783
#>   [29]  978  135  157   24  171  562  994  482  397  649  729  826  701  722
#>   [43]  481  454  697  455  518  197  213  137  255  725   13  287  515  751
#>   [57]  437  743  182  681  357  802   25  886  992  572  698  448  478  203
#>   [71]  451  530  781  772  582  542  300  333  574  869  816  813  849  493
#>   [85]  543  868  261  706  652  631  726  240  570  273  323  483  288  861
#>   [99]   73  818  322  165  727  335  741  488  905  176   67  398  119  538
#>  [113]  524  925  286  297  304  528  800  558  611  829  368  661  367  867
#>  [127]  547   77  852  700  250  579  742  433  334  496  602  637  929  885
#>  [141]  338  198  647  911  639  369   39  917  907  351  896  696  145   74
#>  [155]   75  963  872  599  974  106  101  205  876  469   90   36  926  749
#>  [169]  716  820  759  999  934  646  837   35  301  344  710  150  879   54
#>  [183]  311  274   41  114  497   71  108  416  557   72  485  638   80   92
#>  [197]  230  193  674  280  331  105  719  400  117  841  610  671  739  619
#>  [211]   61  836  184  939  148  702  878  456  349  225  785  955  130  843
#>  [225]  376  927   23  607  928  708  136  682  316  921  343  988  410  680
#>  [239]  822  908  718  471  511  971  859  972  154  282  231  529   81  185
#>  [253]  146  243  393  644   82  356  923  657  835  678  442  441  121  606
#>  [267]  910  158  372  641  226  648  459  975  969  768  565  973  464  774
#>  [281]  431  924  342  465   16  937  133   31   93  811  787  527  935  445
#>  [295]  690  103  113  993  227  388  873  803  411  353  214  775  283  115
#>  [309]  413  695  623  961  110  738  241  209  672  799  998  517  404  655
#>  [323]  216  512  941  952  750  782  257  549  903  390  535  449  685  324
#>  [337]  285  827  332  618  617  856  306  248  594  312  900   52  946  118
#>  [351]  386  575  415  352  714  534  991  210  842  235  788  853  777  600
#>  [365]  476  370  996  328   38  686  234  767  484  409   22  659  981  153
#>  [379]  460  592  462  854  604  143  423  477  104  608   58  378  303  277
#>  [393]  495   62  539  389  138  265  745  596  313  567  169  222  315  247
#>  [407]  723  183  109  408  428  425  590  884  677  540  533  383  218  815
#>  [421]  796  839  124  588  965  149  560  679  276  399  256  458  321   11
#>  [435]  550  901  307  186  116  942  100   65   66   87  345  417  537  246
#>  [449]  736  752  436  863  780  913  346  278  866  341  531  833  757  178
#>  [463]   64  874  401  514  474  424  414  864  251  817  664  951  805  548
#>  [477]  589  967  381  159  694  238  724  847  636  790  758  170  290  427
#>  [491]   47  373  435  944  391  628  147  584  614  865  407  823  653  654
#>  [505]   15  754  438  666  513  337  634  662  221  622  363  580  932  523
#>  [519]    9   95  586  461  597  809  763  765  236  269  174  432  152  888
#>  [533]   88  267  979  180  930   45   76  620  319  731  418   33  970  989
#>  [547]  452  645  909  123  473  997  630  249   98  505  922   86  377  577
#>  [561]  522  825  916  877  429  683   56  340  891  532  472  848  587  958
#>  [575]  500  601  200  207  730  365  375  470  421   63  976  111  746  735
#>  [589]  229  220  233  556  832  195  688  651  982  264  239  254  271  882
#>  [603]  947  561  160  615  880  252  215   44  127  555  943   57  309  957
#>  [617]  289  948  504  151   48  983  632  395  862   43  552  308  466  616
#>  [631]  660  419  684  245  536  732  642  553  450  526  755  224  330  795
#>  [645]  756  860   21  134   10  187   53   94  883  595  175  380  566  895
#>  [659]  670  953  122  640   50  318  797  789  129  292  962  359  320  520
#>  [673]  740  676  384   91  510  857  262  894   12   49  658  987  525  102
#>  [687]  284  571  156  382  317  824  828  228  281  673  379  712  467  715
#>  [701]  598  912  492  168  593   30  851  126  699  144  202  771  360  717
#>  [715]  498  954  223  764  140  920  128  871  675  191  434  141  444  204
#>  [729]  950  663  298  773  188   78  348  519  870  728  327  846  554  840
#>  [743]  737  295  457  591  776   99  211  412   96  959   68  406  621  705
#>  [757]  709   89  889  858  196  420  299  192  426  232  633  499  392  687
#>  [771]  242  162  569   20  573  112  541  563  546  486  296   27  931  583
#>  [785]  949  613  463  667  487  769  890  713  374  669  761  915  402   70
#>  [799]  350  329  506   32  985  762  968  480  179  804   83  272  167  439
#>  [813]  489  721   85  385  430   55  302  387  875  918  263  355  612  791
#>  [827]  933  358  821  665  938  831  812  786  779  516   14  635  887  107
#>  [841]  189  733  190    7  336  509   40  142  521  814  904  838  576  172
#>  [855]  760  181  881  544  850  744   42  793  155  371  120  627  956  199
#>  [869]  711  490  291  326  650  766  748   29   79  691  720  807  131  753
#>  [883]  806  194  977  258  747  906  266  494  559  808  964  902  703  354
#>  [897]  899  403  396  778  794  830  253  984  966  479  693  447  609  260
#>  [911]  314  293   59  259   46  394  568   97  581  219  668  656  453  605
#>  [925]  624  173  206   37  268  551  960  362  310  784  897    8  834  734
#>  [939]  801  643  501  545  361  585   18  940  603  475 1000  844  689  945
#>  [953]  578  201  405  893  339  491  564  707  770   84  364  212  166  629
#>  [967]  792  898  161  704  468  440  132  294  507  275  305  270  208  217
#>  [981]  919  980  139  810  446   28   34  914   51  347  692  366  798  237
#>  [995]  443  503  625  936  125  995